requirements.txt
aiofiles==23.1.0 
aiohttp==3.8.5 
aiosignal==1.3.1 
altair==5.0.1 
anyio==3.7.1 
async-timeout==4.0.2 
attrs==23.1.0 
backoff==2.2.1 
beautifulsoup4==4.12.2 
blinker==1.6.2 
bs4==0.0.1 
cachetools==5.3.1 
certifi==2023.7.22 
cffi==1.15.1 
chardet==5.2.0 
charset-normalizer==3.2.0 
chroma-hnswlib==0.7.2 
chromadb==0.4.5 
click==8.1.6 
colorama==0.4.6 
coloredlogs==15.0.1 
cryptography==41.0.3 
dataclasses-json==0.5.9 
decorator==5.1.1 
et-xmlfile==1.1.0 
exceptiongroup==1.1.2 
fastapi==0.99.1 
fastjsonschema==2.18.0 
filetype==1.2.0 
flatbuffers==23.5.26 
frozenlist==1.4.0 
gitdb==4.0.10 
gitpython==3.1.32 
google-api-core==2.11.1 
google-api-python-client==2.95.0 
google-auth-httplib2==0.1.0 
google-auth==2.22.0 
google-search-results==2.4.2 
googleapis-common-protos==1.60.0 
greenlet==2.0.2 
h11==0.14.0 
httplib2==0.22.0 
httptools==0.6.0 
humanfriendly==10.0 
idna==3.4 
importlib-metadata==6.8.0 
importlib-resources==6.0.0 
install==1.3.5 
jinja2==3.1.2 
joblib==1.3.1 
jsonschema-specifications==2023.7.1 
jsonschema==4.18.6 
jupyter-core==5.3.1 
langchain==0.0.348
langchain-experimental==0.0.45
langsmith==0.0.19 
lxml==4.9.3 
markdown-it-py==3.0.0 
markdown==3.4.4 
markupsafe==2.1.3 
marshmallow-enum==1.5.1 
marshmallow==3.20.1 
mdurl==0.1.2 
monotonic==1.6 
mpmath==1.3.0 
msg-parser==1.2.0 
multidict==6.0.4 
mypy-extensions==1.0.0 
nbformat==5.9.2 
nltk==3.8.1 
numexpr==2.8.4 
numpy==1.25.2 
olefile==0.46 
onnxruntime==1.15.1 
openai==0.27.8 
openapi-schema-pydantic==1.2.4 
openpyxl==3.1.2 
overrides==7.4.0 
packaging==23.1 
pandas==2.0.3 
pdf2image==1.16.3 
pdfminer-six==20221105 
pillow==9.5.0 
pip==23.2.1 
platformdirs==3.10.0 
posthog==3.0.1 
protobuf==4.23.4 
pulsar-client==3.2.0 
pyarrow==12.0.1 
pyasn1-modules==0.3.0 
pyasn1==0.5.0 
pycparser==2.21 
pydantic==1.10.12 
pydeck==0.8.0 
pygments==2.16.1 
pympler==1.0.1 
pypandoc==1.11 
pyparsing==3.1.1 
pypika==0.48.9 
pyreadline3==3.4.1 
python-dateutil==2.8.2 
python-docx==0.8.11 
python-dotenv==1.0.0 
python-magic==0.4.27 
python-pptx==0.6.21 
pytz-deprecation-shim==0.1.0.post0 
pytz==2023.3 
# pywin32==306 
pyyaml==6.0.1 
referencing==0.30.2 
regex==2023.6.3 
requests==2.31.0 
rich==13.5.2 
rpds-py==0.9.2 
rsa==4.9 
semantic-kernel==0.3.5.dev0 
six==1.16.0 
smmap==5.0.0 
sniffio==1.3.0 
soupsieve==2.4.1 
sqlalchemy==2.0.19 
starlette==0.27.0 
streamlit==1.25.0 
sympy==1.12 
tabulate==0.9.0 
tenacity==8.2.2 
tiktoken==0.4.0 
tokenizers==0.13.3 
toml==0.10.2 
toolz==0.12.0 
tornado==6.3.2 
tqdm==4.65.0 
traitlets==5.9.0 
typing-extensions==4.7.1 
typing-inspect==0.9.0 
tzdata==2023.3 
tzlocal==4.3.1 
unstructured==0.8.8 
uritemplate==4.1.1 
urllib3==1.26.16 
uvicorn==0.22.0 
uvicorn[standard]==0.22.0 
uvloop==0.17.0 
validators==0.20.0 
watchdog==3.0.0 
watchfiles==0.19.0 
websockets==11.0.3 
xlrd==2.0.1 
xlsxwriter==3.1.2 
yarl==1.9.2 
zipp==3.16.2 